Move redirect_internal to controllers_frontend.py
* Move `redirect_internal` function from `controllers_auxiliary.py` to `controllers_frontend.py` since it is a frontend controller.
